fix: conditionally add async tag in page editor clientlibs
