+++ src/ModelManager.ts
@@ -17,6 +17,7 @@ import { Model } from './Model';
 import { ModelClient } from './ModelClient';
 import { ModelStore } from './ModelStore';
 import { PathUtils } from './PathUtils';
+import {Utils} from "./Utils";
 
 /**
  * Does the provided model object contains an entry for the given child path
@@ -142,14 +143,12 @@ export class ModelManager {
         }
         // Check if async is required. This will be true only in remote rendering.
         const asyncFlag =  domain === PathUtils.getCurrentPathname() ? false : true;
-        if (asyncFlag && PathUtils.isEditMode()) {
-            const clientLibUrl = this.generateClientLibsUrl(rootModelPath, domain);
-            this.appendClientLibs(clientLibUrl);
+        if (asyncFlag && Utils.isEditMode()) {
+            const clientLibUrl = Utils.generateClientLibsUrl(rootModelPath, domain);
+            Utils.appendClientLibs(clientLibUrl);
         }
         this._editorClient = new EditorClient(this);
         this._modelStore = (initialModel) ? new ModelStore(rootModelPath, initialModel) : new ModelStore(rootModelPath);
-
-
         this._initPromise = this._checkDependencies().then(() => {
 
             const data = this.modelStore.getData(rootModelPath);
@@ -186,33 +185,6 @@ export class ModelManager {
         });
 
         return this._initPromise;
-    }
-
-    /**
-     * Appends client lib to the page
-     */
-    private appendClientLibs(clientLibUrl : string) {
-        const script = document.createElement('script');
-        const scripts = document.getElementsByTagName('script')[0];
-        script.src = clientLibUrl;
-        if (scripts.parentNode) {
-            scripts.parentNode.insertBefore(script, scripts);
-        }
-    }
-
-    /**
-     * Generates clientlib url
-     */
-    private generateClientLibsUrl(pageModelRoot : string, domain : string) : string {
-        let clientLibPath = '';
-        if (pageModelRoot != null) {
-            clientLibPath = domain + '/etc.clientlibs/' +  pageModelRoot.split("/")[2] + '/clientlibs/clientlib-react.min.js';
-        }
-        else {
-            throw new Error('Clientlib path cannot be determined! This should never happen.');
-            return "";
-        }
-        return domain + clientLibPath;
     }
 
     /**
